requirements.txt
absl-py==1.4.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
attrs==23.1.0
awscli==1.27.121
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==6.0.0
boto3==1.26.121
botocore==1.29.121
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
cmdstanpy==0.9.5
colorama==0.4.4
comm==0.1.3
contextlib2==21.6.0
contourpy==1.0.7
convertdate==2.4.0
cycler==0.11.0
Cython==0.29.34
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dlib @ file:///C:/Users/KIIT/Desktop/Dlib-main/dlib-19.22.99-cp38-cp38-win_amd64.whl
docker==6.0.1
docutils==0.16
easyocr==1.7.0
EasyProcess==1.1
ephem==4.1.4
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastjsonschema==2.16.3
filelock==3.12.0
flatbuffers==23.3.3
fonttools==4.39.3
fqdn==1.5.1
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h11==0.14.0
h5py==3.8.0
hijri-converter==2.2.4
holidays==0.23
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==8.2.0
jupyter-console==6.6.3
jupyter-core==5.3.0
jupyter-events==0.6.3
jupyter-server==2.5.0
jupyter-server-terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keras==2.12.0
keyring==8.7
keyrings.alt==4.2.0
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
lazy-loader==0.2
libclang==16.0.0
LunarCalendar==0.0.9
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
ml-dtypes==0.1.0
more-itertools==9.1.0
mpmath==1.3.0
nbclassic==0.5.6
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
ninja==1.11.1
notebook==6.5.4
notebook-shim==0.2.3
numpy==1.24.3
oauthlib==3.2.2
object-detection==0.0.3
opencv-python==4.5.2.54
opencv-python-headless==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.1
pandas==1.2.5
pandocfilters==1.5.0
parso==0.8.3
pdf2image==1.16.3
pickleshare==0.7.5
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
platformdirs==3.4.0
plotly==5.14.1
proglog==0.1.9
prometheus-client==0.16.0
prompt-toolkit==3.0.38
prophet==1.1.2
protobuf==4.22.3
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post4
pycparser==2.21
Pygments==2.15.1
PyMeeus==0.5.12
pynput==1.7.3
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
pystan==2.19.1.1
pytesseract==0.3.8
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
PyVirtualDisplay==2.2
PyWavelets==1.4.1
pywin32==306
pywin32-ctypes==0.2.0
pywinpty==2.0.10
PyYAML==5.4.1
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
requests==2.25.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.7.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
seaborn==0.12.2
selenium==3.10.0
Send2Trash==1.8.2
sentry-sdk==1.9.0
setuptools-git==1.2
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
stack-data==0.6.2
supervision==0.6.0
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.4.12
tinycss2==1.2.1
torch==2.0.0
torchvision==0.15.1
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
typing-extensions==4.5.0
tzdata==2023.3
ultralytics==8.0.88
uri-template==1.2.0
urllib3==1.26.15
utils==1.0.1
voluptuous==0.13.1
wcwidth==0.2.6
webcolors==1.13
webdriver-manager==3.8.6
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.7
wrapt==1.14.1
wsproto==1.2.0
yolo==0.3.1
zipp==3.15.0
